gpu - split common cuda/hip data into separate folder
