+++ backends/cuda-ref/ceed-cuda-operator.c
@@ -22,7 +22,7 @@
 #include <stdbool.h>
 #include <string.h>
 #include "ceed-cuda.h"
-#include "ceed-cuda-compile.h"
+#include "../cuda/ceed-cuda-compile.h"
 
 //------------------------------------------------------------------------------
 // Destroy operator
