+++ backends/cuda-ref/ceed-cuda-qfunction.c
@@ -16,33 +16,32 @@
 
 #include <ceed/ceed.h>
 #include <ceed/backend.h>
-#include <hip/hip_runtime.h>
+#include <cuda.h>
 #include <stdio.h>
 #include <string.h>
-#include "ceed-hip.h"
-#include "ceed-hip-compile.h"
-#include "ceed-hip-qfunction-load.h"
+#include "ceed-cuda.h"
+#include "ceed-cuda-qfunction-load.h"
+#include "../cuda/ceed-cuda-compile.h"
 
 //------------------------------------------------------------------------------
 // Apply QFunction
 //------------------------------------------------------------------------------
-static int CeedQFunctionApply_Hip(CeedQFunction qf, CeedInt Q,
-                                  CeedVector *U, CeedVector *V) {
+static int CeedQFunctionApply_Cuda(CeedQFunction qf, CeedInt Q,
+                                   CeedVector *U, CeedVector *V) {
   int ierr;
   Ceed ceed;
   ierr = CeedQFunctionGetCeed(qf, &ceed); CeedChkBackend(ierr);
 
   // Build and compile kernel, if not done
-  ierr = CeedHipBuildQFunction(qf); CeedChkBackend(ierr);
+  ierr = CeedCudaBuildQFunction(qf); CeedChkBackend(ierr);
 
-  CeedQFunction_Hip *data;
+  CeedQFunction_Cuda *data;
   ierr = CeedQFunctionGetData(qf, &data); CeedChkBackend(ierr);
-  Ceed_Hip *ceed_Hip;
-  ierr = CeedGetData(ceed, &ceed_Hip); CeedChkBackend(ierr);
+  Ceed_Cuda *ceed_Cuda;
+  ierr = CeedGetData(ceed, &ceed_Cuda); CeedChkBackend(ierr);
   CeedInt numinputfields, numoutputfields;
   ierr = CeedQFunctionGetNumArgs(qf, &numinputfields, &numoutputfields);
   CeedChkBackend(ierr);
-  const int blocksize = ceed_Hip->optblocksize;
 
   // Read vectors
   for (CeedInt i = 0; i < numinputfields; i++) {
@@ -64,8 +63,8 @@ static int CeedQFunctionApply_Hip(CeedQFunction qf, CeedInt Q,
 
   // Run kernel
   void *args[] = {&data->d_c, (void *) &Q, &data->fields};
-  ierr = CeedRunKernelHip(ceed, data->qFunction, CeedDivUpInt(Q, blocksize),
-                          blocksize, args); CeedChkBackend(ierr);
+  ierr = CeedRunKernelAutoblockCuda(ceed, data->qFunction, Q, args);
+  CeedChkBackend(ierr);
 
   // Restore vectors
   for (CeedInt i = 0; i < numinputfields; i++) {
@@ -88,31 +87,40 @@ static int CeedQFunctionApply_Hip(CeedQFunction qf, CeedInt Q,
 //------------------------------------------------------------------------------
 // Destroy QFunction
 //------------------------------------------------------------------------------
-static int CeedQFunctionDestroy_Hip(CeedQFunction qf) {
+static int CeedQFunctionDestroy_Cuda(CeedQFunction qf) {
   int ierr;
-  CeedQFunction_Hip *data;
+  CeedQFunction_Cuda *data;
   ierr = CeedQFunctionGetData(qf, &data); CeedChkBackend(ierr);
   Ceed ceed;
   ierr = CeedQFunctionGetCeed(qf, &ceed); CeedChkBackend(ierr);
   if  (data->module)
-    CeedChk_Hip(ceed, hipModuleUnload(data->module));
+    CeedChk_Cu(ceed, cuModuleUnload(data->module));
   ierr = CeedFree(&data); CeedChkBackend(ierr);
+  return CEED_ERROR_SUCCESS;
+}
+
+//------------------------------------------------------------------------------
+// Set User QFunction
+//------------------------------------------------------------------------------
+static int CeedQFunctionSetCUDAUserFunction_Cuda(CeedQFunction qf,
+    CUfunction f) {
+  int ierr;
+  CeedQFunction_Cuda *data;
+  ierr = CeedQFunctionGetData(qf, &data); CeedChkBackend(ierr);
+  data->qFunction = f;
   return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
 // Create QFunction
 //------------------------------------------------------------------------------
-int CeedQFunctionCreate_Hip(CeedQFunction qf) {
+int CeedQFunctionCreate_Cuda(CeedQFunction qf) {
   int ierr;
   Ceed ceed;
   CeedQFunctionGetCeed(qf, &ceed);
-  CeedQFunction_Hip *data;
-  ierr = CeedCalloc(1,&data); CeedChkBackend(ierr);
+  CeedQFunction_Cuda *data;
+  ierr = CeedCalloc(1, &data); CeedChkBackend(ierr);
   ierr = CeedQFunctionSetData(qf, data); CeedChkBackend(ierr);
-  CeedInt numinputfields, numoutputfields;
-  ierr = CeedQFunctionGetNumArgs(qf, &numinputfields, &numoutputfields);
-  CeedChkBackend(ierr);
 
   // Read QFunction source
   ierr = CeedQFunctionGetKernelName(qf, &data->qFunctionName);
@@ -122,9 +130,12 @@ int CeedQFunctionCreate_Hip(CeedQFunction qf) {
 
   // Register backend functions
   ierr = CeedSetBackendFunction(ceed, "QFunction", qf, "Apply",
-                                CeedQFunctionApply_Hip); CeedChkBackend(ierr);
+                                CeedQFunctionApply_Cuda); CeedChkBackend(ierr);
   ierr = CeedSetBackendFunction(ceed, "QFunction", qf, "Destroy",
-                                CeedQFunctionDestroy_Hip); CeedChkBackend(ierr);
+                                CeedQFunctionDestroy_Cuda); CeedChkBackend(ierr);
+  ierr = CeedSetBackendFunction(ceed, "QFunction", qf, "SetCUDAUserFunction",
+                                CeedQFunctionSetCUDAUserFunction_Cuda);
+  CeedChkBackend(ierr);
   return CEED_ERROR_SUCCESS;
 }
 //------------------------------------------------------------------------------
