+++ backends/cuda-ref/ceed-cuda-restriction.c
@@ -16,11 +16,12 @@
 
 #include <ceed/ceed.h>
 #include <ceed/backend.h>
-#include <hip/hip_runtime.h>
+#include <cuda.h>
+#include <cuda_runtime.h>
 #include <stdbool.h>
 #include <stddef.h>
-#include "ceed-hip.h"
-#include "ceed-hip-compile.h"
+#include "ceed-cuda.h"
+#include "../cuda/ceed-cuda-compile.h"
 
 // *INDENT-OFF*
 static const char *restrictionkernels = QUOTE(
@@ -124,21 +125,22 @@ extern "C" __global__ void trOffset(const CeedInt *__restrict__ lvec_indices,
 //------------------------------------------------------------------------------
 // Apply restriction
 //------------------------------------------------------------------------------
-static int CeedElemRestrictionApply_Hip(CeedElemRestriction r,
-                                        CeedTransposeMode tmode, CeedVector u, CeedVector v, CeedRequest *request) {
+static int CeedElemRestrictionApply_Cuda(CeedElemRestriction r,
+    CeedTransposeMode tmode, CeedVector u, CeedVector v, CeedRequest *request) {
   int ierr;
-  CeedElemRestriction_Hip *impl;
+  CeedElemRestriction_Cuda *impl;
   ierr = CeedElemRestrictionGetData(r, &impl); CeedChkBackend(ierr);
   Ceed ceed;
   ierr = CeedElemRestrictionGetCeed(r, &ceed); CeedChkBackend(ierr);
-  Ceed_Hip *data;
+  Ceed_Cuda *data;
   ierr = CeedGetData(ceed, &data); CeedChkBackend(ierr);
-  const CeedInt blocksize = 64;
+  const CeedInt warpsize  = 32;
+  const CeedInt blocksize = warpsize;
   const CeedInt nnodes = impl->nnodes;
   CeedInt nelem, elemsize;
   CeedElemRestrictionGetNumElements(r, &nelem);
   ierr = CeedElemRestrictionGetElementSize(r, &elemsize); CeedChkBackend(ierr);
-  hipFunction_t kernel;
+  CUfunction kernel;
 
   // Get vectors
   const CeedScalar *d_u;
@@ -159,16 +161,16 @@ static int CeedElemRestrictionApply_Hip(CeedElemRestriction r,
       // -- Offsets provided
       kernel = impl->noTrOffset;
       void *args[] = {&nelem, &impl->d_ind, &d_u, &d_v};
-      CeedInt blocksize = elemsize<256?(elemsize>64?elemsize:64):256;
-      ierr = CeedRunKernelHip(ceed, kernel, CeedDivUpInt(nnodes, blocksize),
-                              blocksize, args); CeedChkBackend(ierr);
+      CeedInt blocksize = elemsize<1024?(elemsize>32?elemsize:32):1024;
+      ierr = CeedRunKernelCuda(ceed, kernel, CeedDivUpInt(nnodes, blocksize),
+                               blocksize, args); CeedChkBackend(ierr);
     } else {
       // -- Strided restriction
       kernel = impl->noTrStrided;
       void *args[] = {&nelem, &d_u, &d_v};
-      CeedInt blocksize = elemsize<256?(elemsize>64?elemsize:64):256;
-      ierr = CeedRunKernelHip(ceed, kernel, CeedDivUpInt(nnodes, blocksize),
-                              blocksize, args); CeedChkBackend(ierr);
+      CeedInt blocksize = elemsize<1024?(elemsize>32?elemsize:32):1024;
+      ierr = CeedRunKernelCuda(ceed, kernel, CeedDivUpInt(nnodes, blocksize),
+                               blocksize, args); CeedChkBackend(ierr);
     }
   } else {
     // E-vector -> L-vector
@@ -178,14 +180,14 @@ static int CeedElemRestrictionApply_Hip(CeedElemRestriction r,
       void *args[] = {&impl->d_lvec_indices, &impl->d_tindices,
                       &impl->d_toffsets, &d_u, &d_v
                      };
-      ierr = CeedRunKernelHip(ceed, kernel, CeedDivUpInt(nnodes, blocksize),
-                              blocksize, args); CeedChkBackend(ierr);
+      ierr = CeedRunKernelCuda(ceed, kernel, CeedDivUpInt(nnodes, blocksize),
+                               blocksize, args); CeedChkBackend(ierr);
     } else {
       // -- Strided restriction
       kernel = impl->trStrided;
       void *args[] = {&nelem, &d_u, &d_v};
-      ierr = CeedRunKernelHip(ceed, kernel, CeedDivUpInt(nnodes, blocksize),
-                              blocksize, args); CeedChkBackend(ierr);
+      ierr = CeedRunKernelCuda(ceed, kernel, CeedDivUpInt(nnodes, blocksize),
+                               blocksize, args); CeedChkBackend(ierr);
     }
   }
 
@@ -201,9 +203,9 @@ static int CeedElemRestrictionApply_Hip(CeedElemRestriction r,
 //------------------------------------------------------------------------------
 // Blocked not supported
 //------------------------------------------------------------------------------
-int CeedElemRestrictionApplyBlock_Hip(CeedElemRestriction r, CeedInt block,
-                                      CeedTransposeMode tmode, CeedVector u,
-                                      CeedVector v, CeedRequest *request) {
+int CeedElemRestrictionApplyBlock_Cuda(CeedElemRestriction r, CeedInt block,
+                                       CeedTransposeMode tmode, CeedVector u,
+                                       CeedVector v, CeedRequest *request) {
   // LCOV_EXCL_START
   int ierr;
   Ceed ceed;
@@ -216,10 +218,10 @@ int CeedElemRestrictionApplyBlock_Hip(CeedElemRestriction r, CeedInt block,
 //------------------------------------------------------------------------------
 // Get offsets
 //------------------------------------------------------------------------------
-static int CeedElemRestrictionGetOffsets_Hip(CeedElemRestriction rstr,
+static int CeedElemRestrictionGetOffsets_Cuda(CeedElemRestriction rstr,
     CeedMemType mtype, const CeedInt **offsets) {
   int ierr;
-  CeedElemRestriction_Hip *impl;
+  CeedElemRestriction_Cuda *impl;
   ierr = CeedElemRestrictionGetData(rstr, &impl); CeedChkBackend(ierr);
 
   switch (mtype) {
@@ -236,19 +238,19 @@ static int CeedElemRestrictionGetOffsets_Hip(CeedElemRestriction rstr,
 //------------------------------------------------------------------------------
 // Destroy restriction
 //------------------------------------------------------------------------------
-static int CeedElemRestrictionDestroy_Hip(CeedElemRestriction r) {
+static int CeedElemRestrictionDestroy_Cuda(CeedElemRestriction r) {
   int ierr;
-  CeedElemRestriction_Hip *impl;
+  CeedElemRestriction_Cuda *impl;
   ierr = CeedElemRestrictionGetData(r, &impl); CeedChkBackend(ierr);
 
   Ceed ceed;
   ierr = CeedElemRestrictionGetCeed(r, &ceed); CeedChkBackend(ierr);
-  ierr = hipModuleUnload(impl->module); CeedChk_Hip(ceed, ierr);
+  ierr = cuModuleUnload(impl->module); CeedChk_Cu(ceed, ierr);
   ierr = CeedFree(&impl->h_ind_allocated); CeedChkBackend(ierr);
-  ierr = hipFree(impl->d_ind_allocated); CeedChk_Hip(ceed, ierr);
-  ierr = hipFree(impl->d_toffsets); CeedChk_Hip(ceed, ierr);
-  ierr = hipFree(impl->d_tindices); CeedChk_Hip(ceed, ierr);
-  ierr = hipFree(impl->d_lvec_indices); CeedChk_Hip(ceed, ierr);
+  ierr = cudaFree(impl->d_ind_allocated); CeedChk_Cu(ceed, ierr);
+  ierr = cudaFree(impl->d_toffsets); CeedChk_Cu(ceed, ierr);
+  ierr = cudaFree(impl->d_tindices); CeedChk_Cu(ceed, ierr);
+  ierr = cudaFree(impl->d_lvec_indices); CeedChk_Cu(ceed, ierr);
 
   ierr = CeedFree(&impl); CeedChkBackend(ierr);
   return CEED_ERROR_SUCCESS;
@@ -257,12 +259,12 @@ static int CeedElemRestrictionDestroy_Hip(CeedElemRestriction r) {
 //------------------------------------------------------------------------------
 // Create transpose offsets and indices
 //------------------------------------------------------------------------------
-static int CeedElemRestrictionOffset_Hip(const CeedElemRestriction r,
+static int CeedElemRestrictionOffset_Cuda(const CeedElemRestriction r,
     const CeedInt *indices) {
   int ierr;
   Ceed ceed;
   ierr = CeedElemRestrictionGetCeed(r, &ceed); CeedChkBackend(ierr);
-  CeedElemRestriction_Hip *impl;
+  CeedElemRestriction_Cuda *impl;
   ierr = CeedElemRestrictionGetData(r, &impl); CeedChkBackend(ierr);
   CeedInt nelem, elemsize, lsize, ncomp;
   ierr = CeedElemRestrictionGetNumElements(r, &nelem); CeedChkBackend(ierr);
@@ -321,21 +323,21 @@ static int CeedElemRestrictionOffset_Hip(const CeedElemRestriction r,
 
   // Copy data to device
   // -- L-vector indices
-  ierr = hipMalloc((void **)&impl->d_lvec_indices, nnodes*sizeof(CeedInt));
-  CeedChk_Hip(ceed, ierr);
-  ierr = hipMemcpy(impl->d_lvec_indices, lvec_indices,
-                   nnodes*sizeof(CeedInt), hipMemcpyHostToDevice);
-  CeedChk_Hip(ceed, ierr);
+  ierr = cudaMalloc((void **)&impl->d_lvec_indices, nnodes*sizeof(CeedInt));
+  CeedChk_Cu(ceed, ierr);
+  ierr = cudaMemcpy(impl->d_lvec_indices, lvec_indices,
+                    nnodes*sizeof(CeedInt), cudaMemcpyHostToDevice);
+  CeedChk_Cu(ceed, ierr);
   // -- Transpose offsets
-  ierr = hipMalloc((void **)&impl->d_toffsets, sizeOffsets*sizeof(CeedInt));
-  CeedChk_Hip(ceed, ierr);
-  ierr = hipMemcpy(impl->d_toffsets, toffsets, sizeOffsets*sizeof(CeedInt),
-                   hipMemcpyHostToDevice); CeedChk_Hip(ceed, ierr);
+  ierr = cudaMalloc((void **)&impl->d_toffsets, sizeOffsets*sizeof(CeedInt));
+  CeedChk_Cu(ceed, ierr);
+  ierr = cudaMemcpy(impl->d_toffsets, toffsets, sizeOffsets*sizeof(CeedInt),
+                    cudaMemcpyHostToDevice); CeedChk_Cu(ceed, ierr);
   // -- Transpose indices
-  ierr = hipMalloc((void **)&impl->d_tindices, sizeIndices*sizeof(CeedInt));
-  CeedChk_Hip(ceed, ierr);
-  ierr = hipMemcpy(impl->d_tindices, tindices, sizeIndices*sizeof(CeedInt),
-                   hipMemcpyHostToDevice); CeedChk_Hip(ceed, ierr);
+  ierr = cudaMalloc((void **)&impl->d_tindices, sizeIndices*sizeof(CeedInt));
+  CeedChk_Cu(ceed, ierr);
+  ierr = cudaMemcpy(impl->d_tindices, tindices, sizeIndices*sizeof(CeedInt),
+                    cudaMemcpyHostToDevice); CeedChk_Cu(ceed, ierr);
 
   // Cleanup
   ierr = CeedFree(&ind_to_offset); CeedChkBackend(ierr);
@@ -348,13 +350,13 @@ static int CeedElemRestrictionOffset_Hip(const CeedElemRestriction r,
 //------------------------------------------------------------------------------
 // Create restriction
 //------------------------------------------------------------------------------
-int CeedElemRestrictionCreate_Hip(CeedMemType mtype, CeedCopyMode cmode,
-                                  const CeedInt *indices,
-                                  CeedElemRestriction r) {
+int CeedElemRestrictionCreate_Cuda(CeedMemType mtype, CeedCopyMode cmode,
+                                   const CeedInt *indices,
+                                   CeedElemRestriction r) {
   int ierr;
   Ceed ceed;
   ierr = CeedElemRestrictionGetCeed(r, &ceed); CeedChkBackend(ierr);
-  CeedElemRestriction_Hip *impl;
+  CeedElemRestriction_Cuda *impl;
   ierr = CeedCalloc(1, &impl); CeedChkBackend(ierr);
   CeedInt nelem, ncomp, elemsize;
   ierr = CeedElemRestrictionGetNumElements(r, &nelem); CeedChkBackend(ierr);
@@ -403,24 +405,24 @@ int CeedElemRestrictionCreate_Hip(CeedMemType mtype, CeedCopyMode cmode,
       break;
     }
     if (indices != NULL) {
-      ierr = hipMalloc( (void **)&impl->d_ind, size * sizeof(CeedInt));
-      CeedChk_Hip(ceed, ierr);
+      ierr = cudaMalloc( (void **)&impl->d_ind, size * sizeof(CeedInt));
+      CeedChk_Cu(ceed, ierr);
       impl->d_ind_allocated = impl->d_ind; // We own the device memory
-      ierr = hipMemcpy(impl->d_ind, indices, size * sizeof(CeedInt),
-                       hipMemcpyHostToDevice);
-      CeedChk_Hip(ceed, ierr);
-      ierr = CeedElemRestrictionOffset_Hip(r, indices); CeedChkBackend(ierr);
+      ierr = cudaMemcpy(impl->d_ind, indices, size * sizeof(CeedInt),
+                        cudaMemcpyHostToDevice);
+      CeedChk_Cu(ceed, ierr);
+      ierr = CeedElemRestrictionOffset_Cuda(r, indices); CeedChkBackend(ierr);
     }
   } else if (mtype == CEED_MEM_DEVICE) {
     switch (cmode) {
     case CEED_COPY_VALUES:
       if (indices != NULL) {
-        ierr = hipMalloc( (void **)&impl->d_ind, size * sizeof(CeedInt));
-        CeedChk_Hip(ceed, ierr);
+        ierr = cudaMalloc( (void **)&impl->d_ind, size * sizeof(CeedInt));
+        CeedChk_Cu(ceed, ierr);
         impl->d_ind_allocated = impl->d_ind; // We own the device memory
-        ierr = hipMemcpy(impl->d_ind, indices, size * sizeof(CeedInt),
-                         hipMemcpyDeviceToDevice);
-        CeedChk_Hip(ceed, ierr);
+        ierr = cudaMemcpy(impl->d_ind, indices, size * sizeof(CeedInt),
+                          cudaMemcpyDeviceToDevice);
+        CeedChk_Cu(ceed, ierr);
       }
       break;
     case CEED_OWN_POINTER:
@@ -431,7 +433,7 @@ int CeedElemRestrictionCreate_Hip(CeedMemType mtype, CeedCopyMode cmode,
       impl->d_ind = (CeedInt *)indices;
     }
     if (indices != NULL) {
-      ierr = CeedElemRestrictionOffset_Hip(r, indices); CeedChkBackend(ierr);
+      ierr = CeedElemRestrictionOffset_Cuda(r, indices); CeedChkBackend(ierr);
     }
   } else {
     // LCOV_EXCL_START
@@ -440,38 +442,38 @@ int CeedElemRestrictionCreate_Hip(CeedMemType mtype, CeedCopyMode cmode,
     // LCOV_EXCL_STOP
   }
 
-  // Compile HIP kernels
+  // Compile CUDA kernels
   CeedInt nnodes = impl->nnodes;
-  ierr = CeedCompileHip(ceed, restrictionkernels, &impl->module, 8,
-                        "RESTRICTION_ELEMSIZE", elemsize,
-                        "RESTRICTION_NELEM", nelem,
-                        "RESTRICTION_NCOMP", ncomp,
-                        "RESTRICTION_NNODES", nnodes,
-                        "RESTRICTION_COMPSTRIDE", compstride,
-                        "STRIDE_NODES", strides[0],
-                        "STRIDE_COMP", strides[1],
-                        "STRIDE_ELEM", strides[2]); CeedChkBackend(ierr);
-  ierr = CeedGetKernelHip(ceed, impl->module, "noTrStrided",
-                          &impl->noTrStrided); CeedChkBackend(ierr);
-  ierr = CeedGetKernelHip(ceed, impl->module, "noTrOffset", &impl->noTrOffset);
+  ierr = CeedCompileCuda(ceed, restrictionkernels, &impl->module, 8,
+                         "RESTRICTION_ELEMSIZE", elemsize,
+                         "RESTRICTION_NELEM", nelem,
+                         "RESTRICTION_NCOMP", ncomp,
+                         "RESTRICTION_NNODES", nnodes,
+                         "RESTRICTION_COMPSTRIDE", compstride,
+                         "STRIDE_NODES", strides[0],
+                         "STRIDE_COMP", strides[1],
+                         "STRIDE_ELEM", strides[2]); CeedChkBackend(ierr);
+  ierr = CeedGetKernelCuda(ceed, impl->module, "noTrStrided",
+                           &impl->noTrStrided); CeedChkBackend(ierr);
+  ierr = CeedGetKernelCuda(ceed, impl->module, "noTrOffset", &impl->noTrOffset);
   CeedChkBackend(ierr);
-  ierr = CeedGetKernelHip(ceed, impl->module, "trStrided", &impl->trStrided);
+  ierr = CeedGetKernelCuda(ceed, impl->module, "trStrided", &impl->trStrided);
   CeedChkBackend(ierr);
-  ierr = CeedGetKernelHip(ceed, impl->module, "trOffset", &impl->trOffset);
+  ierr = CeedGetKernelCuda(ceed, impl->module, "trOffset", &impl->trOffset);
   CeedChkBackend(ierr);
 
   // Register backend functions
   ierr = CeedSetBackendFunction(ceed, "ElemRestriction", r, "Apply",
-                                CeedElemRestrictionApply_Hip);
+                                CeedElemRestrictionApply_Cuda);
   CeedChkBackend(ierr);
   ierr = CeedSetBackendFunction(ceed, "ElemRestriction", r, "ApplyBlock",
-                                CeedElemRestrictionApplyBlock_Hip);
+                                CeedElemRestrictionApplyBlock_Cuda);
   CeedChkBackend(ierr);
   ierr = CeedSetBackendFunction(ceed, "ElemRestriction", r, "GetOffsets",
-                                CeedElemRestrictionGetOffsets_Hip);
+                                CeedElemRestrictionGetOffsets_Cuda);
   CeedChkBackend(ierr);
   ierr = CeedSetBackendFunction(ceed, "ElemRestriction", r, "Destroy",
-                                CeedElemRestrictionDestroy_Hip);
+                                CeedElemRestrictionDestroy_Cuda);
   CeedChkBackend(ierr);
   return CEED_ERROR_SUCCESS;
 }
@@ -479,7 +481,7 @@ int CeedElemRestrictionCreate_Hip(CeedMemType mtype, CeedCopyMode cmode,
 //------------------------------------------------------------------------------
 // Blocked not supported
 //------------------------------------------------------------------------------
-int CeedElemRestrictionCreateBlocked_Hip(const CeedMemType mtype,
+int CeedElemRestrictionCreateBlocked_Cuda(const CeedMemType mtype,
     const CeedCopyMode cmode, const CeedInt *indices, CeedElemRestriction r) {
   int ierr;
   Ceed ceed;
