+++ backends/cuda-shared/ceed-cuda-shared.c
@@ -18,7 +18,6 @@
 #include <ceed/backend.h>
 #include <string.h>
 #include "ceed-cuda-shared.h"
-#include "../cuda/ceed-cuda.h"
 
 //------------------------------------------------------------------------------
 // Backend init
