+++ backends/hip-ref/ceed-hip-operator.c
@@ -21,7 +21,7 @@
 #include <stdbool.h>
 #include <string.h>
 #include "ceed-hip.h"
-#include "ceed-hip-compile.h"
+#include "../hip/ceed-hip-compile.h"
 
 //------------------------------------------------------------------------------
 // Destroy operator
