+++ backends/hip-ref/ceed-hip-vector.c
