+++ backends/hip-shared/ceed-hip-shared-basis.c
@@ -19,7 +19,7 @@
 #include <hip/hip_runtime.h>
 #include <stddef.h>
 #include "ceed-hip-shared.h"
-#include "../hip/ceed-hip.h"
+#include "../hip/ceed-hip-common.h"
 #include "../hip/ceed-hip-compile.h"
 
 //------------------------------------------------------------------------------
