+++ backends/hip/ceed-hip-common.c
@@ -17,46 +17,47 @@
 #include <ceed/ceed.h>
 #include <ceed/backend.h>
 #include <string.h>
-#include "ceed-hip-gen.h"
+#include <stdlib.h>
+#include "ceed-hip-common.h"
 
 //------------------------------------------------------------------------------
-// Backend init
+// Device information backend init
 //------------------------------------------------------------------------------
-static int CeedInit_Hip_gen(const char *resource, Ceed ceed) {
+int CeedHipInit(Ceed ceed, const char *resource, int nrc) {
   int ierr;
-  const int nrc = 8; // number of characters in resource
-  if (strncmp(resource, "/gpu/hip/gen", nrc))
-    // LCOV_EXCL_START
-    return CeedError(ceed, CEED_ERROR_BACKEND,
-                     "Hip backend cannot use resource: %s", resource);
-  // LCOV_EXCL_STOP
+  const char *device_spec = strstr(resource, ":device_id=");
+  const int deviceID = (device_spec) ? atoi(device_spec+11) : -1;
+
+  int currentDeviceID;
+  ierr = hipGetDevice(&currentDeviceID); CeedChk_Hip(ceed,ierr);
+  if (deviceID >= 0 && currentDeviceID != deviceID) {
+    ierr = hipSetDevice(deviceID); CeedChk_Hip(ceed,ierr);
+    currentDeviceID = deviceID;
+  }
+
+  struct hipDeviceProp_t deviceProp;
+  ierr = hipGetDeviceProperties(&deviceProp, currentDeviceID);
+  CeedChk_Hip(ceed,ierr);
 
   Ceed_Hip *data;
-  ierr = CeedCalloc(1, &data); CeedChkBackend(ierr);
-  ierr = CeedSetData(ceed, data); CeedChkBackend(ierr);
-  ierr = CeedHipInit(ceed, resource, nrc); CeedChkBackend(ierr);
-
-  Ceed ceedshared;
-  CeedInit("/gpu/hip/shared", &ceedshared);
-  ierr = CeedSetDelegate(ceed, ceedshared); CeedChkBackend(ierr);
-
-  const char fallbackresource[] = "/gpu/hip/ref";
-  ierr = CeedSetOperatorFallbackResource(ceed, fallbackresource);
-  CeedChkBackend(ierr);
-
-  ierr = CeedSetBackendFunction(ceed, "Ceed", ceed, "QFunctionCreate",
-                                CeedQFunctionCreate_Hip_gen); CeedChkBackend(ierr);
-  ierr = CeedSetBackendFunction(ceed, "Ceed", ceed, "OperatorCreate",
-                                CeedOperatorCreate_Hip_gen); CeedChkBackend(ierr);
-  ierr = CeedSetBackendFunction(ceed, "Ceed", ceed, "Destroy",
-                                CeedDestroy_Hip); CeedChkBackend(ierr);
+  ierr = CeedGetData(ceed, &data); CeedChkBackend(ierr);
+  data->deviceId = currentDeviceID;
+  data->optblocksize = 256;
   return CEED_ERROR_SUCCESS;
 }
 
 //------------------------------------------------------------------------------
-// Register backend
+// Backend Destroy
 //------------------------------------------------------------------------------
-CEED_INTERN int CeedRegister_Hip_Gen(void) {
-  return CeedRegister("/gpu/hip/gen", CeedInit_Hip_gen, 20);
+int CeedDestroy_Hip(Ceed ceed) {
+  int ierr;
+  Ceed_Hip *data;
+  ierr = CeedGetData(ceed, &data); CeedChkBackend(ierr);
+  if (data->hipblasHandle) {
+    ierr = hipblasDestroy(data->hipblasHandle); CeedChk_Hipblas(ceed, ierr);
+  }
+  ierr = CeedFree(&data); CeedChkBackend(ierr);
+  return CEED_ERROR_SUCCESS;
 }
+
 //------------------------------------------------------------------------------
